// FrontEnd/frontend.client/src/app/interfaces/Contrato.ts
export interface Contrato {
  id: number
  numero: string
  objeto: string
  nombre: string
  empresaId: number
  nombreContratista: string
  numeroDocumentoContratista: string
  correoContratista: string
  telefonoContratista: string
  valor: number
  duracion: number
  fechaInicio: string
  fechaFin: string
}